# perf: wait-free WAL