https://github.com/influxdata/influxdb_iox/issues/437