@CJP10 2020-11-12
@alamb Sorry I didn't make it clear, this is still quite WIP. I'm not sure what y'alls convention is for WIP but this would definitely qualify as WIP.

 I'll fix up the tests whens it more feature complete. I'll also do my best to get [loom](https://github.com/tokio-rs/loom) working to verify some concurrency edge cases.

In terms of hard performance numbers, I don't have those. The impl isn't done yet so I'd have to cross that bridge first. I'd also have to write some benchmarks for both the new impl and old impl. I didn't see any benches for the old impl but I could have missed them. I'm currently operating under  some assumptions/educated guesses. One of those being that using the parallelism of the entire runtime and not using channels to middleman the syscalls will be a positive impact on performance.

I guess I can go over a bit of the impl and hope it does justice or sparks some more questions.

The reason for using `crossbeam_epoch` to wrap the WAL files is to avoid a use-after-free post rollover.
Drilling down a bit, when threads call append and pass the `rollover_size` they race to mark the `active_file` ptr.
The thread that marks the pointer is responsible for creating the new WAL file and CAS the `active_file` with the one created.
`last_active_file` exists because a task can be preempted during a rollover and come back to execute after the rollover.
The thread will then use `last_active_file` to write that payload. This isn't perfect because the task can be preempted indefinitely and so the correct approach is to store a list of every WAL file so you write back to any time. `ROLLOVER_SEQUENCE_NUMBER_BUFFER` also plays a part as it allows further writing to a WAL file after rollover. The current value is somewhat arbitrary but finding a sensible default seems straight forward. Regardless, this is needed because the exact rollover point can't be determined so we leave room for outstanding writes to the previous WAL file.

Int terms of safety, a majority of the unsafe is documented on the crossbeam_epoch side but I wouldn't mind putting a comment block on each giving some more detail.