@alamb 2020-11-12
> Sorry I didn't make it clear, this is still quite WIP. I'm not sure what y'alls convention is for WIP but this would definitely qualify as WIP.

@CJP10  Cool -- that makes sense -- sorry I was confused. 

Also, thank you for the explanation