@CJP10 2020-11-12
@alamb No worries! I'll refine the impl a bit more and when it's ready I'll shoot you a ping if that's cool.