@CJP10 2020-11-15
@alamb there is still some plumbing left to do but the overall impl of `append` is pretty much done. If you could give it a read and let me know what you think?