@CJP10 2020-11-18
> In terms of the design, I wonder if you have any benchmarks or usecases that were driving this implementation? As in, I am sure this design is better is several aspects than the current code, but it does have drawbacks too (like the use of `unsafe`)?

On my machine (Ryzen 5950x@Stock and Samsung 980 Pro 500GB) I was able to get ~4700MB/s (~5000MB/s is what I got just using fio to benchmark the disk). So it's pretty close to running at raw disk speed without fsyncing. I don't think you can get match faster than the current append hot path. I came up with this when implementing a K-V store based on [WiscKey](https://www.usenix.org/system/files/conference/fast16/fast16-papers-lu.pdf). Unfortunately, `unsafe` is the nature of the beast and can't really be avoided in this impl.

> Also, can you explain why you removed the existing `wal/tests` test files? The scenarios in them seem to still be relevant to this implementation. Also I didn't see many (any?) tests for the new code that was added.

I took some inspirations from the existing tests but there wasn't much carry over. The file rollover test didn't apply because file rollover isn't deterministic and so verifying the exact details of the files wasn't of value. The delete up to test wouldn't provide much value because it's just deleting a file. I wouldn't mind adding back the old tests but the new tests do cover most of the edge cases (located at the bottom of the lib.rs).


> Also, can you explain what guarantees this implementation gives compared to the current WAL (which upon completion of `write_and_sync` guarantees the data has been written to the wal and `fsync` has returned, giving some additional durability to the data). I probably don't understand fully, but it seems like this implementation purposely doesn't wait for fsync, which probably does increase performance but at the expense of durability.

Added syncing :smile:.