@alamb 2020-11-18
Thanks @CJP10 

> On my machine (Ryzen 5950x@Stock and Samsung 980 Pro 500GB) I was able to get ~4700MB/s (~5000MB/s is what I got just using fio to benchmark the disk

Is it possible to share the benchmark / scripts you were using? Did you compare this PR's performance to the performance on `main`? 

The comparison that is probably the most relevant is when `fsync` is enabled