@alamb 2020-11-18
The other thing I wonder as I spend time reading about crossbeam/epoch is if you tested this implementation under concurrent load (aka when there are  multiple writers across different threads), aka in the type of scenario that I would expect unsafe behavior to potentially be tricky to get right

Our project does not yet have this type of tooling / benchmark, so I don't have any specific suggested commands