@CJP10 2020-11-18
> Thanks @CJP10
> 
> > On my machine (Ryzen 5950x@Stock and Samsung 980 Pro 500GB) I was able to get ~4700MB/s (~5000MB/s is what I got just using fio to benchmark the disk
> 
> Is it possible to share the benchmark / scripts you were using? Did you compare this PR's performance to the performance on `main`?
> 
> The comparison that is probably the most relevant is when `fsync` is enabled

I didn't compare perf but I can commit my benches and port them to main for the existing WAL.

> In general, I don't understand is how this code ensures the WAL data isn't corrupted by multiple concurrent writers. I probably don't understand something about the design (see comment on `append`)
> 
> I was trying to play with this code locally, but it didn't compile for me (on my MacOS laptop):
> 
> ```
>    Compiling generated_types v0.1.0 (/Users/alamb/Software/influxdb_iox2/generated_types)
> error[E0432]: unresolved import `nix::sys::uio::pwritev`
>  --> wal/src/io.rs:1:21
>   |
> 1 | use nix::sys::uio::{pwritev, IoVec};
>   |                     ^^^^^^^
>   |                     |
>   |                     no `pwritev` in `sys::uio`
>   |                     help: a similar name exists in the module: `writev`
> 
> error[E0433]: failed to resolve: use of undeclared type `SeekFrom`
>   --> wal/src/io.rs:56:15
>    |
> 56 |     file.seek(SeekFrom::Start(offset)).context(FailedToSeek)?;
>    |               ^^^^^^^^ not found in this scope
>    |
> ```

That's my bad I have to cfg that dep anyway on non-supported platforms.