@alamb 2020-11-22
#TLDR
It is my opinion this PR doesn't improve performance sufficiently to justify the additional complexity introduced (in terms of using `unsafe`, having os specific compilation flags, and additional dependencies).

Maybe I am missing something, but I don't think we should take this change given what I have seen.

I would be happy to contemplate a WAL change that reduced the calls to write using safe code. @CJP10  we sort of failed on this PR as we don't yet have clear contribution guidelines here that would have made the decision making process more transparent and could have focused us more on performance benchmarks)


# Discussion:
I compared the performance of this branch against the WAL on master using the wal bench provided (along with variants that wrote the same data 10 and 100 times)

While I saw a 30% improvement on this branch compared to master, I believe almost all of that is attributable to using fewer system IO calls on each write (the current WAL implementation on main does a few using unbuffered writes). When I simply added buffering into the WAL on main it goes  5 times faster than either implementation.

Not that I am advocating using buffered IO in the WAL, but I am using that observation as additional evidence that the WAL is (unsurprisingly) entirely IO bound; Allowing multiple threads to write concurrently (as in this PR) does not change the fact it is IO bound nor change its performance characteristics.

Here are some performance results using  my macbook pro

# This PR

Using the WAL code in this PR + some fixes to get it to compile on mac: https://github.com/influxdata/influxdb_iox/tree/alamb/wal-with-bench @ a91e96ea1fcc28335ea0852e9ff8c4d5536163dd

```

wal/append-1            time:   [8.9782 us 9.0758 us 9.2008 us]
                        thrpt:  [424.55 MiB/s 430.40 MiB/s 435.08 MiB/s]


wal/append-10           time:   [89.346 us 90.176 us 91.093 us]
                        thrpt:  [428.82 MiB/s 433.18 MiB/s 437.20 MiB/s]


wal/append-100          time:   [880.35 us 887.56 us 896.55 us]
                        thrpt:  [435.70 MiB/s 440.11 MiB/s 443.72 MiB/s]
````

# Main

Against the WAL on main (without calls to `sync_all` so I think it is a reasonably fair comparison): https://github.com/influxdata/influxdb_iox/tree/alamb/wal-perf-on-main @ 0999adbe2663bd9d4704b6ec775db2e0e54b3364

```
wal/append-1            time:   [10.649 us 10.736 us 10.827 us]
                        thrpt:  [360.79 MiB/s 363.85 MiB/s 366.80 MiB/s]

wal/append-10           time:   [107.82 us 110.23 us 113.80 us]
                        thrpt:  [343.27 MiB/s 354.36 MiB/s 362.30 MiB/s]

wal/append-100          time:   [1.0710 ms 1.0795 ms 1.0884 ms]
                        thrpt:  [358.91 MiB/s 361.85 MiB/s 364.73 MiB/s]
```


# Main using buffered IO

When using a buffered writer (which isn't actually a good idea, but since we aren't sync'ig the writes anyways, I don't think it changes the durability story materially):

On https://github.com/influxdata/influxdb_iox/tree/alamb/wal-with-buffering at 4af4ac67ba0df90d61f2411ddd6f23a09c15d309


```
wal/append-1            time:   [1.6421 us 1.6674 us 1.7033 us]
                        thrpt:  [2.2396 GiB/s 2.2878 GiB/s 2.3231 GiB/s]

wal/append-10           time:   [16.681 us 17.064 us 17.731 us]
                        thrpt:  [2.1514 GiB/s 2.2355 GiB/s 2.2868 GiB/s]

wal/append-100          time:   [167.25 us 168.83 us 170.52 us]
                        thrpt:  [2.2371 GiB/s 2.2595 GiB/s 2.2808 GiB/s]
```

cc @CJP10